requirements.txt
# algoliasearch==2.6.2
# algoliasearch-django==2.0.0
arabic-reshaper==2.1.4
asgiref==3.5.2
asn1crypto==1.5.1
attrs==22.1.0
autopep8==1.7.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
cssselect2==0.7.0
defusedxml==0.7.1
# dj-database-url==1.0.0
dj-rest-auth==2.2.5
Django == 3.2
django-allauth==0.50.0

django-cors-headers==3.10.1
django-filter==22.1
django-heroku==0.3.1
django-rest-auth==0.9.5
django-rest-enumfield==0.2.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-spectacular==0.24.2
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.3
importlib-resources==5.9.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.16.0
lxml==4.9.1
MarkupSafe==2.1.1
oauthlib==3.2.0
openapi-codec==1.3.2
oscrypto==1.3.0
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
psycopg2==2.9.3
pycodestyle==2.9.1
pycparser==2.21
pyHanko==0.15.1
pyhanko-certvalidator==0.19.6
PyJWT==2.4.0
PyPDF3==1.0.6
pyrsistent==0.18.1
python-bidi==0.4.2
python3-openid==3.2.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.12
requests==2.28.1
requests-oauthlib==1.3.1
simplejson==3.17.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
stripe==4.2.0
svglib==1.4.1
tinycss2==1.2.1
toml==0.10.2
tqdm==4.64.1
tzdata==2022.6
tzlocal==4.2
uritemplate==4.1.1
uritools==4.0.0
urllib3==1.26.12
webencodings==0.5.1
whitenoise==5.3.0
xhtml2pdf==0.2.8
zipp==3.8.1
